+++ intelmq/bots/outputs/restapi/output.py
@@ -16,6 +16,11 @@ class RestAPIOutputBot(Bot):
 
         self.session = requests.Session()
         self.set_request_parameters()
+        self.session.proxies.update(self.proxy)
+        self.session.headers.update(self.http_header)
+        self.session.verify = self.http_verify_cert
+        self.session.cert = self.ssl_client_cert
+
         if self.parameters.auth_token_name and self.parameters.auth_token:
             if self.parameters.auth_type == 'http_header':
                 self.session.headers.update(
@@ -34,10 +39,6 @@ class RestAPIOutputBot(Bot):
             kwargs = {'data': event.to_dict(hierarchical=self.parameters.hierarchical_output)}
 
         r = self.session.post(self.parameters.host,
-                              proxies=self.proxy,
-                              headers=self.http_header,
-                              verify=self.http_verify_cert,
-                              cert=self.ssl_client_cert,
                               timeout=self.http_timeout_sec,
                               **kwargs)
         if not r.ok: